Fix KUserGroup::users() and KUserGroup::userNames() on UNIX
The gr_mem member of struct group does not contain the names of users
where the primary group is equal to that group. In order to fix this we
have to iterate over all users and check whether the primary gid matches
the requested group. Unlike getgrouplist() there does not seem to be a
function that performs this task for us.

REVIEW: 116883