Remove KPluginLoader dependency on KService
The reason for this is so that KPluginLoader, KPluginFactory and
kexportplugin.h do not require anything else from the KService
framework, allowing them to be moved to KCoreAddons.

A KPluginName class is introduced to act as an intermediary between
KService and KPluginLoader.  The KPluginLoader(KService) constructor is
replaced with KPluginLoader(KPluginName), and KService gains a cast
operator to convert it to a KPluginName.

Now
  KService service("blah.desktop");
  KPluginLoader(service);
works as before, maintaining source compatibility.

The reason for the intermediary class (rather than just, say, providing
a QString() operator to KService) is to make sure this combination is
the only one that works (and you can't start using KService everywhere a
QString is expected).

(REVIEW): 117185

Commit 953f6afab51bc2249be9cd73ff1c59b0092afdea in kservice