Implement KUser::faceIconPath on Windows.
I use an undocumented Windows API that stores the profile image in a
temporary file and returns the path to it. The previous code was just
trying to load that temporary file, and would only work if another app
had created the file recently (such as the control panel section where
the image is changed).

This only works on Windows Vista and later; on Windows XP the undocumented
API is different, and faceIconPath will just return an empty string.

REVIEW:117131