Deprecate the catalog name stuff from KAboutData
This is pretty useless - the translation catalog has to be set before
KAboutData is constructed in order to translate its arguments.

Note that we cannot provide a constructor overload that is identical
other than omitting the catalog name argument, as that would be
ambiguous (if 4 strings are passed, it could be either the new
constructor or the old one).  So instead, we provide two new
constructors: one requiring at least 4 strings and the license type, and
one only taking 3 strings.

REVIEW: 117275