Fix unit tests for KDirWatch when TMPDIR's filesystem reuses inode numbers.
When the filesystem TMPDIR points to (by default /tmp) reuses inode numbers,
the testDeleteAndRecreateFile test will fail as neither the stat nor the
QFileSystemWatcher backends can detect the deletion/recreation as two discrete
events.  They only detect the fact the file changed, and emit the dirty signal.
Fix the test suite to check for this case, and allow the test to proceed.

REVIEW: 118612