When using QFileSystemWatcher, try to avoid having deleted files stop being watched.
When a file is deleted, QFileSystemWatcher does not explicitly signal this fact,
instead emitting a generic changed signal.  However, it will stop watching the file
in question, at least on Linux with the inotify backend.  Normally this is ok, but
if a file is quickly recreated and the inode number is unchanged, KDirWatch assumes
the file is still monitored and no further changes to the file will occur.

Instead, KDirWatch will now always try to re-add the file to QFileSystemWatcher.
It doesn't bother checking if the operation succeeded, as in most cases it will
fail as the file is still watched.  It can only assume the file is now being
watched at this point.

REVIEW: 118612