KUser: Fix build on MSVC2012, working around a compiler bug.
Trying to convert a unique_ptr<T, someLambda> to bool gives an "ambiguous
conversion" error in MSVC2012. Worked around it by using an explicit
function object instead of a lambda.

This causes a second error, in both 2012 and 2013 ("getting confused
between the global and the injected FooDeleter") which is worked around
by using ::FooDeleter instead of FooDeleter.

BUG:337997