Add command line option to update cache
This will, once finished, generate an index file in the plugins
directory which holds the metadata of the plugins. The index file will
be used from KPluginLoader / KPluginTrader for faster queries.

The data will be saved in json binary format. As we also want to
generate indices from .desktop files, this code needs to go here,
otherwise we end up copying the parser bits elsewhere (all the
interesting things are private).

This is only the skeleton, implementation comes next.