Declare InheritanceChecker before actual use
This is actually not needed for "proper" compilers, because the use is inside a template,
The type is only required at instantiation time.

However, this patch makes the Coverity build tool happy. Without this patch,
we get an error for every translation unit including
kpluginfactory.h, telling us that "InheritanceChecker is not a
template"

Also see https://communities.coverity.com/thread/2903

REVIEW: 120435