decide at build time where the home of KDE4 is
Kdelibs4Migration used some heuristic to decide where ~/.kde4 was, was simply checking ~/.kde then ~/.kde4.
but in Plasma5 we need to be really sure where ~/.kde is since some settings are kept in sync even at srtartup (where ~/.kde doesn't exist, and we have to create it from kf5, that's why the kdehome() method)
also, checking for ~/.kde and ~/.kde4 may break in cases where a distribution uses ~/.kde4 for kde4, and used to have ~/.kde for kde3 (remnants of kde3 configs in user homes is very possible)

this adds a cmake option _KDE4_DEFAULT_HOME_POSTFIX coherent with _KDE_DEFAULT_HOME_POSTFIX of kdelibs4. so it would be responsibility of distributions to make this right.

REVIEW:120480