Introduce KPluginMetaData::metaDataSource()
When creating a KPluginMetaData object from a .desktop file we currently
set fileName() to point to the .desktop file. However, passing
fileName() to KPluginLoader should allow loading the plugin. To fix this
we store the value of the X-KDE-Library key as fileName() instead and
move the path to the .desktop file to a new property metaDataSource()

REVIEW: 121313