Add KTextToHTML class for plaintext->HTML conversion
This class comes from KPimUtils::LinkLocator from kdepimlibs. It takes
plaintext string, and converts it to HTML while perserving spacing, converting
URLs and email addresses into <a href...> links and interpreting highligting
markup (bold, underline, /italic/) and wrapping it to corresponding HTML tags.
It can also convert text emoticons into <img> smileys.

Since smileys conversion depends on KEmoticons, it cannot be in KCoreAddons directly.
The feature it's implemented through a plugin in KEmoticons. When the KEmoticons
framework is not available, or the code is invoked without QGuiApplication, the
smileys conversion will return the text unaltered.

REVIEW: 121094