Fix build of ktexttohtmltest on Windows.
ktexttohtmltest has ktexttohtml.cpp as a source file *and* it's linking to
the kcoreaddons library which includes the same source file, so it defines
the same function. In addition, ktexttohtml.h in this context is declaring
the function as __declspec(dllimport), so it expects to find the function
in a linked library. All this causes a multiple definition error when
linking the unit test.

Fixed by not linking to the KCoreAddons library in the unit test. Set
include directories such that kcoreaddons_export is found (previously it
was being found because of linking to the KF5::CoreAddons CMake target),
and fake static linking so that the export macros do nothing (to avoid the
"inconsistent linkage" warning).