Always return absolute paths in Kdelibs4Migration
KDELibs4Migration returns an absolute path in the case that .kde or
.kde4 exist.

If that folder does not exist and we fall back to KDE4_DEFAULT_HOME we
also want an absolute path.

This fixes the kde4breeze script which is launched from kconf_update
writing a kdeglobals file in the wrong directory.

REVIEW: 121595