Add support for Hidden key
This key does not make a lot of sense for binary plugins (those loaded
from .so files which retrieve their metadata from the json-embedded
data), but is useful for plugins loaded from .desktop files.
This allows us to override the visibility of installed plugins by the
user.

I'm only considering the Hidden key here. I am not sure how the
NoDisplay key (which seems to be used here and there), so maybe we
should also consider NoDisplay. This is the most basic patch we can
achieve the desired results with, however.

REVIEW:122619