Add KPluginMetaData::formFactor API
This is a new global key in the json metadata (or .desktop files) that returns a
stringlist of formfactors this plugin or application is useful for. It
may return any string in the list (we don't know yet which formfactors
may be useful in the future).

The usecase here is that some applications may want to load certain
plugins only on a given formfactor, or apps or plugins that should be
started with a different command line argument depending on the
formfactor. Also, app launcher or plugin listers may filter the list of
offered/shown plugins by this FormFactor key.

CHANGELOG:Allow filtering plugins and applications by formfactor
REVIEW:124066