KDirWatch: fix thread safety
The use of KDirWatch from KSycoca exposed that KDirWatch was really not
thread-safe. The internal singleton design brings a ton of races between
KDirWatch instances. Easy solution: one "singleton" per thread, using
QThreadStorage.

The main impact is on memory management; the "singleton" is no longer refcounted
but simply deleted at the end of the thread, by QThreadStorage (including for
the main thread). Since this happens at ~QApp time, we need to zero the d pointer
of KDirWatch::self(), which is deleted by a global-static dtor later on.

REVIEW: 125217