Add KPluginMetaData::translators() and KAboutPerson::fromJson()
Plugins could have a list of translators in KDE4 through KAboutData so
this commit restores this for KF5. As a side effect of now having
authors() and translators() returning a list of KAboutPerson I also
added KAboutPerson::fromJSON() as public function but I can make that
internal API instead if it's preferred.

REVIEW: 126343