Add function to get runtime frameworks version information
Adds a method similar to qVersion() that returns a string of the
frameworks version being run. This differs from the header file which
can only provide the version this app is compiled against.

A numerical version for runtime comparions is also provided.

FEATURE: Add API to get Frameworks runtime version info
REVIEW: 127031