[kshareddatacache] Fix invalid use of & to avoid unaligned reads
That the & is wrong here is obvious:
align can never be 2 inside the if condition as (2 & bool) is always false,
yet the switch on align contains a case for 2.
Instead it drops into the while loop for an 4-byte aligned pointer.

This doesn't change the returned values on platforms where unaligned
reads are permitted, so I didn't bump the version.

REVIEW: 127968
CCMAIL: fabian@ritter-vogt.de