First stab at indexed plugin loading
This only works for absolute paths, but seems to work quite nicely at
that.

One problem with this version is that the creation of the
KPluginMetaData and the directory lookup are somewhat shielded,
API-wise. We want to hook into right before the diriterator is used,
and keep our cache in memory at least during that loop, but the metadata
object is created from the filename (which means it's going through the
more expensive, file-system-touching ctor), while we want to create it
from the JSON object. We can't just change the signature of
forEachPlugin(...) to pass the metadata, since that's public API.