DesktopFileParser: Honor ServiceTypes field
In case there's a ServiceType field in the desktop file, look them up for
definitions.

Otherwise we just printed a warning whenever the cmake macro was used that
nobody resolved (see plasma workspace).

This patch assumes some heuristic for the naming that is not standardized
(/ to -, all lower-case), it won't enforce it though, so worst case
scenario it works just as it does now.

REVIEW: 129643