kformat: Make it possible to properly translate relative dates.
KFormat::formatRelativeDate allows one to obtain a human-readable date like
"Last Wednesday" or "Next Thursday" for dates within a week of the current
date.

The returned string cannot be properly translated in all languages since it's
generated using a placeholder of the form "Next %1" -- the word Next (or Last)
may itself need to be modified to be correct which can't be done with the
current translation system.

The fix is just to return "Next $dayName" directly, based on the day name as
appropriate, and allow the translators to translate the entire phrase.

The auto tests are adjusted as well. They were actually incorrect before this
change, since the API documentation for formatRelativeDate makes clear that the
date format is only applied if the date is *not* within a week of the current
date.

BUG:335106
FIXED-IN:5.31

Differential Revision: https://phabricator.kde.org/D4023