Do not watch QRC's paths
Watching a QRC path is not supported and results in `"."` being
watched which can lead to problems.

For example in bug #374075 KIO adds ":/kio5/newfile-templates"
as path to watch (this is probably another bug in itself).
If we are already watching "/home/user", this breaks the emission
of the dirty() signal for every new children of "/home/user" (somehow,
the relative path is used for them, e.g. "./foo.txt" instead of
"/home/user/foo.txt"). In particular, in inotifyEventReceived()
e->m_client is empty and so e->path is not added to
e->m_pendingFileChanges. This only happens if "/home/user" is also
the cwd of the process using KDirWatch.

Ignoring QRC paths fixes this issue.

BUG: 374075
FIXED-IN: 5.35

Test Plan:
From dolphin, Create New -> Text File in a folder which is also the current working
directory of the dolphin process.

Reviewers: dfaure

Differential Revision: https://phabricator.kde.org/D5877