add KAboutLicense::spdx and introduce orLater qualification
Summary:
External software (e.g. appstream) uses the standardized SPDX license
identifiers. Seeing as they are specified and ours are not it is useful to
have a quick way to convert our licenses to SPDX notation.

To also preserve 'this version or later versions' notations properly so
we can reflect them into SPDX there now is an orLater bool which tracks
the qualification. The qualification is both for us and for SPDX denoted
as a trailing plus sign. To that end we'll infer that input keywords with
a trailing plus sign are orLater. For code-constructed licenses via
kAboutData.addLicense() there is a new explicit parameter orLater to
qualify the license. The param defaults to false to retain backwards
compatibility.
Constructing KAboutData with orLater set is not supported as the param list
is long enough.

::spdx() returns expressions. Right now they are simple ones (because we
actually do not support anything more involved from a KAboutLicense POV)
such as 'GPL-2.0+'. In the future this can easily be
'GPL-2.0+ WITH excpetion' though if we decide to support that.

Private::spdxID() is internally used to translate our enum key to the
well-known identifier used by SPDX. It does not construct an expression.

CHANGELOG: New spdx API on KAboutLicense to get SPDX license expressions

Test Plan: - new test passes

Reviewers: sebas, mpyne

Reviewed By: mpyne

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D6672