desktoptojson: Improve legacy service type detection heuristic.
This commit adds a second-pass heuristic converting requests for named
service types into possible file names for corresponding Desktop Entry
files as used in previous KDE releases.  The current heuristic converts
'/' in the requested service type name to '-'.  If that fails to find a
file, we now also convert '/' to '' (i.e. nothing) and try again before
returning an error.

This fix introduces an autotest regression if you have KDevelop
installed, since the updated code now finds the *real* KDevelop/Plugin
service type desktop entry in a test where it was expecting not to, the
"kdevcppnolanguagesupport no servicetype" test.  To avoid installed
software affecting the tests I rename the service type and verify that
the "no servicetype" and "with [fake] servicetype" tests continue to
pass.

BUG:384037
FIXED-IN:5.39

Differential Revision: https://phabricator.kde.org/D8002