Verify that the file was opened
Summary:
The assertion is disabled in release mode, but the verify will show
up even in a release build. In debug mode, the test won't abort
anymore, but this shouldn't happen anyways. If there's some other
problem, like a full disk, then the error is still visible and we
can check what's going on.

Reviewers: dfaure

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D9820