Increase timeout in the hope of fixing test failure on the CI
The benchmark works fine for me on my local machine, but on the
CI the QFileSystemWatcher based benchmark fails. I suspect that
it is due to a timing issue. Let's see if that hunch is correct
by increasing the timeout.

Note: Profiling that test shows lots of low-hanging fruit in
the QFileSystemWatcher implementation. Something we should fix
upstream, notably lots of QStringList accesses which could
probably be fixed by using QSet<QString> or similar.

CCMAIL: dfaure@kde.org