Bump shared-mime-info to 1.3
Summary:
Bump the build time requirement, and assumed runtime, version of
shared-mime-info to 1.3, which seems to be a good low-enough baseline
according to:
https://repology.org/metapackage/shared-mime-info/versions

This makes it possible to drop a number of definitions provided by
newer versions of shared-mime-info:
- application/metalink+xml
- application/x-mimearchive
- application/x-mobipocket-ebook
- application/x-nzb
- application/x-trig
- audio/aac
- image/webp

Also, drop text/x-objchdr too, as it has no magic nor glob, hence it is
basically useless.

Test Plan: Builds fine.

Reviewers: dfaure

Reviewed By: dfaure

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D10569