Remove misleading use of KDE_INSTALL_PYTHONBINDINGSDIR
Summary:
Although KDE_INSTALL_PYTHONBINDINGSDIR follows the the standard pattern
of KDE_INSTALL_<dir> for installation directories, it is not included
in the set of directories set by KDEInstallDirs.cmake.

INSTALL_DIR_SUFFIX does not specify a path but just a suffix and defaults
to "site-packages", i.e. the suffix used by some (but not all) distros.

Removing the (typically empty) optional argument from the function
invocation avoids confusion and allows to remove the parameter.

Test Plan: make

Reviewers: #frameworks, apol

Reviewed By: apol

Subscribers: apol, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D15558