[KAutoSaveFile] Improve the comment in tempFileName()
Summary:
Hopefully now it's clear that the filename length in this context is the
filename component only, e.g. test.txt, and not the length of the whole
path, e.g. /home/timon/test.txt. (I think part of the confusion probably
comes from the fact that from the POV of a filesystem a dir is a file with
special attributes...).

Reviewers: #frameworks, mpyne, dfaure

Reviewed By: dfaure

Subscribers: mardelle, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D25979