demote plugin load errors from warning to debug level + reword
Summary:
while they are errors from our POV the callers POV may be that
they'll simply try another name. with the previous behavior
we would always end up printing the message because warnings
are visible by default and then on top of that the warning
would use the word error.

this lead to users getting confused about the actual impact:
https://bugs.kde.org/show_bug.cgi?id=364637

when the caller has additional handling:
https://phabricator.kde.org/D26791

the latter ought to always be the case as far as error reporting
goes because the class documentation even suggests as much
in its example snippet.

as a result of all this it seems way more appropriate to
consider this a debug message on a kcoreaddons level.
additionally, the message now talks about failure to load rather
than error loading so as to make it a bit less important
sounding.

Test Plan: -

Reviewers: mpyne, ngraham

Reviewed By: mpyne, ngraham

Subscribers: ngraham, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D26839