KPluginMetaData: handle invalid mimetype in supportsMimeType
QMimeDatabase doesn't check for validity in inherits, so oktetapart
was found even for invalid mimetypes (because application/octet-stream
is the implicit parent of everything).

Not adding a qCWarning because this is called for every single part
available on the system, from *one* application call with an invalid
mimetype.

Found while writing a KParts unittest.