Allow formatting of values that are not expressed in the base unit
Often times when dealing with data coming from external sources it can already
be in a prefixed unit. Currently one has to be careful to not make mistakes
converting the value to the base unit before passing it to KFormat. This is
adds two methods that take care of this.