Add *.kcrash glob pattern to KCrash Report MIME type
Now that Bug 209041 is fixed, the temporary DrKonqi workaround
http://commits.kde.org/kde-runtime/c4a72e847f4d3461f76205a0b732648d752063b8
is going to be reverted. Therefore the restored default extension should
be recognized as KCrash Report. Let us keep *.kcrash.txt glob pattern as
well for the time being to be compatible with older DrKonqi versions and
to recognize crash report files created in the past.