WIP: [kpluginmetadata] Add metadata entry for plugin's config module
Some plugins have config modules associated with them. KPluginInfo exposes that via the kcmServices method. KPluginMetaData does not have any equivalent.

I think the way KPluginInfo implments it is overly complicated. It uses KServiceTypeTrader and the ParentComponent metadata to guess what plugin a config module might belong to.

It's much simpler and straight forward to have the plugin itself define the id of its config module instead of the other way around.

This patch adds a new metadata property of type string that is the id of the config module. It therefore assumes that a plugin has at most one config module,
which sounds like a reasonable assumption to me.