KJob: add setProgressUnit(), to choose how percent is calculated
Ever since this class was written, there was a d->progressUnit member
variable, but the setter was missing.

I'm going to need this in CopyJob when renaming files, we don't have the
file sizes since they don't matter, so we can calculate progress based
on the number of files renamed instead.