Add KStringHandler::isSubsequence
Can be used to determine whether a given pattern is a subsequence
of another string, for instance "libremath" would match "LibreOffice Math",
useful for e.g. search and filter features.

This is a port of KApplicationTrader::isSubsequence since it really
belongs into KStringHandler which makes it more accessible by reducing
its dependencies, of which there are none on KService anyway.