kfuzzymatcher: Use simpler best score output scheme.
This partially restores the behavior of the code which handles updating
the output parameter holding the "best fuzzy score" to the version prior
to commit 7a212acd.

I did this because I believe it is possible for the modification to fail
to copy all of the matches into the output when the score found by
a recursive call ends up matching a longer subsequence than the match
found by the calling function (whose match goes up to `nextMatch`).

There may be a logical reason that this code is correct (and therefore
more optimal), but based on the impending tagging date I prefer to use
the safer version for now until we validate correctness here.