kpluginfactory: Use the user-provided base factory in our factory macro.
The existing macro always inherits from KPluginFactory -- making it hard
to use when sub-classing KPluginFactory in order to extend it.  The
parameter baseFactory to the macro isn't used.

Use baseFactory as the base class for the plugin.

BUG:410851

-- Commentary by committer:

This closes out merge request !1. Due to the length of time between when
the MR was authored and when we could finally land it, this also
includes other commits added as part of the MR to:

* Use @-style Doxygen comments consistently
* Adapt other Doxygen API documentation to our style guide
* Some other miscellaneous editorial improvements

On top of those changes, I have ensured that the other changes to this
header made in the meantime now conform to the consistent API
documentation (i.e. other \ to @ conversions), and to update the @since
version to be accurate again.