FreeBSD: use procstat again
In 115e003281f a check was added to restrict procstat to just the
BSD systems; however,

	string(REGEX MATCH "[Bb][Ss][Dd]" BSDLIKE CMAKE_SYSTEM_NAME)

this does not expand the variable CMAKE_SYSTEM_NAME: it checks
that **as a string** against the regex, which obviously doesn't
match because "bsd" does not occur in the variable name.

Use ${CMAKE_SYSTEM_NAME} instead, to expand to "FreeBSD" (or
whatever), which **will** match.