KJob: if a delegate is already attached to a job, warn when setJob() is called
Trying to attach a job to an already occupied UI delegate failed silently,
instead show a warning so that the caller fix the code. This could happen
if you try to attach the same UI delegate of a parent KCompositeJob to one
of its sub-jobs.

Also make the code in KJob::setUiDelegate() more readable by splitting the
logic.