Network mounts performance optimizations
This patch is part of performance optimizations for mounted network shares,
see also kio, kconfigwidgets, kwidgetsaddons, ktexteditor, kate, dolphin.

It adds a class KNetworkMounts which can be used to conveniently
query and set performance settings.

In KFileSystemType::fileSystemType it allows to determine certain paths to
return filesystem type KFileSystemType::Smb or KFileSystemType::Nfs without
an expensive call to statfs().

In KDirWatchPrivate::addWatch INotify can be selected to be used on slow
paths.

KDirWatch::addDir and KDirWatch::addFile can be skipped on slow paths.