Fix kcoreaddons_desktop_to_json when cross-compiling
We should not be using the service types from the host, but instead we
should be looking for them in ${CMAKE_SYSROOT}. We now reports an error
at CMake time if the service type cannot be found instead of when running
desktoptojson. Found while cross-compiling kwin for FreeBSD on a macOS
host since it failed to find kcmodule.desktop before this change.