KFileSystemType: minor improvements; detect NTFS filesystem
The central location of superblock magic numbers on Linux seems to be
<linux/magic.h>.

grep'ing in in /usr/src/<kernel-version>/fs/ we can find more superblock
MAGIC numbers. They're not in any public header, so just document where I
found them.

AUTOFSNG_SUPER_MAGIC I couldn't find any references of, so left it as-is.