Downgrade the "invalid plugin" warning message to a debug message
The KPluginMetadata constructor is used to probe for valid plugins so this
results in lots of warning messages being printed when starting e.g.
Kate or plasmashell. As this is expected behaviour we should not be
emitting warning messages which suggest that something is wrong.

See discussion on https://invent.kde.org/frameworks/kcoreaddons/-/merge_requests/112