KPluginFactory: Provide more context for warning
Now the warning says which object is responsible for the issue. Otherwise one can only manually debug
it or guess. This is for example the case in the systemsettings default indicators where lots of plugins are loaded.

The error message now says:
```
kf.coreaddons: Two plugins with the same interface (QObject) were registered in the KPluginFactory KcmComponentChooserFactory. This might be due to a missing Q_OBJECT macro in one of the registered classes
```

The registering using keywords is also not the proper fix in this case and will
be removed in KF6, see https://phabricator.kde.org/T14744