Write value of QT_MAJOR_VERSION in cmake config file
In case the apps don't include ECM first & don't have this var manually set,
the variable evaluates to an empty sting.
Because we require the Qt version of the project to be the same as the one KCoreAddons
was compiled with, it does not make sense for consumers to be able to change the QT_MAJOR_VERSION value.

This is for example the case for heaptrack.