KPluginMetadata: store all paths as absolute ones
We were storing some of the paths as absolute paths (e.g. for
QPluginLoader-based metadata), but in some cases we were just storing
the path unmodified. For consistency this commit normalizes all paths
that were used for opening a file and adds a test to verify that all
paths are absolute. Path arguments that have not been used for reading
are still stored as passed in.