Remove *.doc pattern for text/plain
Long ago, the intent was to avoid misdetecting text files in /usr/share
as msword files. But this change wasn't accepted upstream in
shared-mime-info so we did it locally.

These days, I see no reason to diverge from upstream (better have the
same results everywhere), my /usr/share has no *.doc files anymore
(in all of /usr I can only find /usr/lib64/python2.7/pdb.doc and
/usr/share/kbd/keymaps/i386/qwerty/no-latin1.doc)

And if someone really really cares, then they can file a bug report
for shared-mime-info, it's not our job to diverge on this.

And this was breaking kdelibs4support's kmimetypetest in the new CI
because our kde5.xml is installed in a different prefix than
shared-mime-info so it fully takes precedence instead of "adding" to it.